Update README.md
I hope this works, this is my first time using github and was the only way I found to submit a request:) I added some comments in between to help clarify as well. Sorry I won't be able to join you guys today, but very excited to have found time again to be working on this! Feel free to email/skype/gchat, etc. with questions/requests. 

I'm going to squeeze in some time now for the other things we talked about as well.